Make imatrix not a header-only lib
